requirements-dev.txt
# Development dependencies for PGSD

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Code quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.5.0
isort>=5.12.0

# Security
safety>=2.3.0
bandit[toml]>=1.7.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Build and packaging
build>=0.10.0
twine>=4.0.0

# Development utilities
pre-commit>=3.3.0
tox>=4.6.0

# Database testing (optional)
psycopg2-binary>=2.9.0; sys_platform != "win32"
psycopg2>=2.9.0; sys_platform == "win32"

# YAML parsing
PyYAML>=6.0

# Structured logging
structlog>=23.0.0